+++ src/parse.rs
@@ -1,0 +1,43 @@
+use crate::structure::{Atom, Crystal};
+use anyhow::{Context, Result};
+// Function to parse XYZ file format from string content
+fn parse_xyz_content(contents: &str) -> Result<Crystal> {
+    let lines = contents.lines().collect::<Vec<&str>>();
+
+    if lines.len() < 2 {
+        return Err(anyhow::anyhow!("XYZ file too short"));
+    }
+
+    // First line should contain the number of atoms
+    let num_atoms: usize = lines[0]
+        .trim()
+        .parse()
+        .context("Failed to parse number of atoms")?;
+
+    // Second line is a comment (we can skip it)
+    // Remaining lines contain atom data
+
+    let mut atoms = Vec::new();
+
+    for (i, line) in lines.iter().skip(2).enumerate() {
+        if i >= num_atoms {
+            break;
+        }
+
+        let parts: Vec<&str> = line.split_whitespace().collect();
+        if parts.len() < 4 {
+            continue; // Skip malformed lines
+        }
+
+        let atom = Atom {
+            element: parts[0].to_string(),
+            x: parts[1].parse().context("Failed to parse x coordinate")?,
+            y: parts[2].parse().context("Failed to parse y coordinate")?,
+            z: parts[3].parse().context("Failed to parse z coordinate")?,
+        };
+
+        atoms.push(atom);
+    }
+
+    Ok(Crystal { atoms })
+}
